chore: refactor date options
